# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI IT Helpdesk Agent
  description: AI IT Helpdesk Agent backend with chat, ticketing, and HR request APIs.
  version: 0.1.0
paths:
  /api/v1/chat/message:
    post:
      tags:
      - Chat
      summary: Send a message to the helpdesk agent
      description: Accept a chat message and return the agent response.
      operationId: send_message_api_v1_chat_message_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/chat/stream:
    get:
      tags:
      - Chat
      summary: Stream assistant response as server-sent events
      description: Stream token-by-token chat responses for frontend consumption over
        SSE.
      operationId: stream_message_api_v1_chat_stream_get
      parameters:
      - name: message
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Message
      - name: session_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Session Id
      - name: employee_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Authenticate a user and issue a JWT
      description: Authenticate a user with email and password.
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Return the current authenticated user
      description: Return the active user profile for the current request.
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/auth/logout:
    post:
      tags:
      - Authentication
      summary: Invalidate the current session
      description: Log the current user out of the active session.
      operationId: logout_api_v1_auth_logout_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/password-reset/initiate:
    post:
      tags:
      - Password Reset
      summary: Request a secure password reset
      description: Start the password reset flow for a user account.
      operationId: initiate_api_v1_password_reset_initiate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/password-reset/verify:
    post:
      tags:
      - Password Reset
      summary: Verify a password reset token
      description: Finalize a password reset using the supplied token.
      operationId: verify_api_v1_password_reset_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/leave/balance/{employee_id}:
    get:
      tags:
      - Leave
      summary: Get the current leave balance
      description: Return the employee leave balance.
      operationId: balance_api_v1_leave_balance__employee_id__get
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveBalanceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/leave/history/{employee_id}:
    get:
      tags:
      - Leave
      summary: Get leave history
      description: Return leave history for an employee.
      operationId: history_api_v1_leave_history__employee_id__get
      parameters:
      - name: employee_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Employee Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaveHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AuthLoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 8
          title: Password
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: AuthLoginRequest
    AuthTokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          const: bearer
          title: Token Type
          default: bearer
      additionalProperties: false
      type: object
      required:
      - access_token
      title: AuthTokenResponse
    ChatMessageRequest:
      properties:
        session_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Session Id
        message:
          type: string
          maxLength: 10000
          minLength: 1
          title: Message
        employee_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
      additionalProperties: false
      type: object
      required:
      - message
      title: ChatMessageRequest
    ChatMessageResponse:
      properties:
        session_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Session Id
        response:
          type: string
          title: Response
        intent:
          type: string
          title: Intent
          default: unknown
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          default: 0.0
        citations:
          items:
            type: string
          type: array
          title: Citations
        sources:
          items:
            $ref: '#/components/schemas/SourceDocument'
          type: array
          title: Sources
      additionalProperties: false
      type: object
      required:
      - response
      title: ChatMessageResponse
    CurrentUserResponse:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        email:
          type: string
          format: email
          title: Email
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        roles:
          items:
            type: string
          type: array
          title: Roles
        is_active:
          type: boolean
          title: Is Active
          default: true
      additionalProperties: false
      type: object
      required:
      - user_id
      - email
      title: CurrentUserResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LeaveBalanceResponse:
      properties:
        employee_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
        casual_leave_days:
          type: number
          minimum: 0.0
          title: Casual Leave Days
        sick_leave_days:
          type: number
          minimum: 0.0
          title: Sick Leave Days
        earned_leave_days:
          type: number
          minimum: 0.0
          title: Earned Leave Days
        carried_over_days:
          type: number
          minimum: 0.0
          title: Carried Over Days
          default: 0
        as_of:
          type: string
          format: date
          title: As Of
      additionalProperties: false
      type: object
      required:
      - casual_leave_days
      - sick_leave_days
      - earned_leave_days
      - as_of
      title: LeaveBalanceResponse
    LeaveHistoryItemResponse:
      properties:
        history_id:
          type: string
          format: uuid
          title: History Id
        leave_type:
          type: string
          enum:
          - casual
          - sick
          - earned
          - unpaid
          title: Leave Type
        start_date:
          type: string
          format: date
          title: Start Date
        end_date:
          type: string
          format: date
          title: End Date
        days:
          type: number
          title: Days
        status:
          type: string
          enum:
          - pending
          - approved
          - rejected
          - cancelled
          title: Status
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      additionalProperties: false
      type: object
      required:
      - history_id
      - leave_type
      - start_date
      - end_date
      - days
      - status
      title: LeaveHistoryItemResponse
    LeaveHistoryResponse:
      properties:
        employee_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employee Id
        history:
          items:
            $ref: '#/components/schemas/LeaveHistoryItemResponse'
          type: array
          title: History
      additionalProperties: false
      type: object
      title: LeaveHistoryResponse
    MessageResponse:
      properties:
        message:
          type: string
          title: Message
      additionalProperties: false
      type: object
      required:
      - message
      title: MessageResponse
    PasswordResetConfirmRequest:
      properties:
        token:
          type: string
          minLength: 1
          title: Token
        new_password:
          type: string
          minLength: 12
          title: New Password
      additionalProperties: false
      type: object
      required:
      - token
      - new_password
      title: PasswordResetConfirmRequest
    PasswordResetRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
      additionalProperties: false
      type: object
      required:
      - email
      title: PasswordResetRequest
    PasswordResetResponse:
      properties:
        message:
          type: string
          title: Message
      additionalProperties: false
      type: object
      required:
      - message
      title: PasswordResetResponse
...
    SourceDocument:
      properties:
        source_filename:
          type: string
          title: Source Filename
        category:
          type: string
          title: Category
        last_updated:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Last Updated
        header:
          anyOf:
          - type: string
          - type: 'null'
          title: Header
        page_content:
          anyOf:
          - type: string
          - type: 'null'
          title: Page Content
      additionalProperties: false
      type: object
      required:
      - source_filename
      - category
      title: SourceDocument
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...